Move logging-related code to mod log
